requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
Pillow>=9.5.0
tqdm>=4.65.0
tensorboard>=2.13.0
open-clip-torch>=2.20.0

